Begin splitting localisation view into subcomponents
