+++ nusight2/src/client/components/localisation/entities/field/view.tsx
