+++ nusight2/src/client/components/localisation/entities/grid/view.tsx
