+++ nusight2/src/client/components/localisation/entities/nugus_robot/config/nugus.glb
