+++ nusight2/src/client/components/localisation/entities/nugus_robot/stories/model_visualizer.tsx
