+++ nusight2/src/client/components/localisation/entities/nugus_robot/stories/nugus_robot.stories.tsx
