+++ nusight2/src/client/components/localisation/entities/skybox/skybox.vert
