+++ nusight2/src/client/components/localisation/model.ts
@@ -5,9 +5,9 @@ import { Vector3 } from "../../../shared/math/vector3";
 import { memoize } from "../../base/memoize";
 import { AppModel } from "../app/model";
 
-import { FieldModel } from "./field/model";
+import { FieldModel } from "./entities/field/model";
 import { LocalisationRobotModel } from "./robot_model";
-import { SkyboxModel } from "./skybox/model";
+import { SkyboxModel } from "./entities/skybox/model";
 
 export class TimeModel {
   @observable time: number; // seconds
