+++ nusight2/src/client/components/localisation/view.tsx
@@ -18,14 +18,17 @@ import { PerspectiveCamera } from "../three/three_fiber";
 import { ThreeFiber } from "../three/three_fiber";
 
 import { LocalisationController } from "./controller";
-import { FieldView } from "./field/view";
+import { FieldView } from "./entities/field/view";
 import roboto from "./fonts/Roboto Medium_Regular.json";
-import { GridView } from "./grid/view";
+import { GridView } from "./entities/grid/view";
 import { LocalisationModel } from "./model";
 import { ViewMode } from "./model";
 import { LocalisationNetwork } from "./network";
 import { LocalisationRobotModel } from "./robot_model";
-import { SkyboxView } from "./skybox/view";
+import { SkyboxView } from "./entities/skybox/view";
+
+import { LocalisationViewModel } from "./view_model";
+
 type LocalisationViewProps = {
   controller: LocalisationController;
   Menu: ComponentType<{}>;
@@ -61,11 +64,10 @@ export class FieldDimensionSelector extends React.Component<FieldDimensionSelect
           {FieldDimensionOptions.map((option) => (
             <div
               key={option.value}
-              className={`flex p-2 ${
-                this.props.model.field.fieldType === option.value
-                  ? "hover:bg-auto-contrast-1"
-                  : "hover:bg-auto-contrast-1"
-              }`}
+              className={`flex p-2 ${this.props.model.field.fieldType === option.value
+                ? "hover:bg-auto-contrast-1"
+                : "hover:bg-auto-contrast-1"
+                }`}
               onClick={() => this.props.controller.setFieldDimensions(option.value, this.props.model)}
             >
               <Icon size={24}>
@@ -324,632 +326,3 @@ function viewModeString(viewMode: ViewMode) {
       throw new Error(`No string defined for view mode ${viewMode}`);
   }
 }
-
-export const LocalisationViewModel = observer(({ model }: { model: LocalisationModel }) => {
-  return (
-    <object3D>
-      <PerspectiveCamera
-        args={[75, 1, 0.01, 100]}
-        position={model.camera.position.toArray()}
-        rotation={[Math.PI / 2 + model.camera.pitch, 0, -Math.PI / 2 + model.camera.yaw, "ZXY"]}
-        up={[0, 0, 1]}
-      >
-        <pointLight color="white" />
-      </PerspectiveCamera>
-      <SkyboxView model={model.skybox} />
-      <hemisphereLight args={["#fff", "#fff", 0.6]} />
-      {model.fieldVisible && <FieldView model={model.field} />}
-      {model.gridVisible && <GridView />}
-      {model.robotVisible &&
-        model.robots
-          .filter((robotModel) => robotModel.visible)
-          .map((robotModel) => <Robot key={robotModel.id} model={robotModel} />)}
-      {model.fieldLinePointsVisible && <FieldLinePoints model={model} />}
-      {model.ballVisible && <Balls model={model} />}
-      {model.fieldIntersectionsVisible && <FieldIntersections model={model} />}
-      {model.particlesVisible && <Particles model={model} />}
-      {model.goalVisible && <Goals model={model} />}
-      {model.walkToDebugVisible &&
-        model.robots
-          .filter((robot) => robot.visible && robot.Hfd)
-          .map((robot) => <WalkPathVisualiser key={robot.id} model={robot} />)}
-      {model.robots
-        .filter((robot) => robot.visible && robot.Hft && robot.purpose)
-        .map((robot) => (
-          <PurposeLabel
-            key={robot.id}
-            robotModel={robot}
-            cameraPitch={model.camera.pitch}
-            cameraYaw={model.camera.yaw}
-          />
-        ))}
-      {model.walkToDebugVisible &&
-        model.robots
-          .filter((robot) => robot.visible && robot.Hfd)
-          .map((robot) => <WalkPathGoal key={robot.id} model={robot} />)}
-      <Robots model={model} />
-      {model.boundedBoxVisible &&
-        model.robots.map((robot) => {
-          if (robot.visible && robot.boundingBox) {
-            return <BoundingBox key={robot.id} model={robot} />;
-          }
-          return null;
-        })}
-    </object3D>
-  );
-});
-
-const WalkPathVisualiser = ({ model }: { model: LocalisationRobotModel }) => {
-  if (!model.Hfd || !model.Hfr) {
-    return null;
-  }
-  const rDFf = model.Hfd?.decompose().translation;
-  const rTFf = model.Hft.decompose().translation;
-  const robot_rotation = new THREE.Euler().setFromQuaternion(model.Hft.decompose().rotation.toThree(), "XYZ");
-  const target_rotation = new THREE.Euler().setFromQuaternion(model.Hfd.decompose().rotation.toThree(), "XYZ");
-  const min_align_radius = model.min_align_radius;
-  const max_align_radius = model.max_align_radius;
-  const min_angle_error = model.min_angle_error;
-  const max_angle_error = model.max_angle_error;
-  const angle_to_final_heading = model.angle_to_final_heading;
-  const Rfr = new THREE.Quaternion(
-    model.Hfr?.decompose().rotation.x,
-    model.Hfr?.decompose().rotation.y,
-    model.Hfr?.decompose().rotation.z,
-    model.Hfr?.decompose().rotation.w,
-  );
-  const vRf = model.velocity_target.toThree().applyQuaternion(Rfr);
-
-  const velocity_direction = Math.atan2(vRf.y, vRf.x);
-  const speed = Math.sqrt(vRf.x ** 2 + vRf.y ** 2) * 1.5;
-
-  const arrowGeometry = (length: number) => {
-    const arrowShape = new THREE.Shape();
-
-    arrowShape.moveTo(0, -0.01);
-    arrowShape.lineTo(0, 0.01);
-    arrowShape.lineTo(length * 0.7, 0.01);
-    arrowShape.lineTo(length * 0.7, 0.02);
-    arrowShape.lineTo(length, 0);
-    arrowShape.lineTo(length * 0.7, -0.02);
-    arrowShape.lineTo(length * 0.7, -0.01);
-    arrowShape.lineTo(0, -0.01);
-
-    const geometry = new THREE.ShapeGeometry(arrowShape);
-
-    return geometry;
-  };
-
-  return (
-    <object3D>
-      <mesh position={[rDFf?.x, rDFf?.y, 0.005]}>
-        <circleBufferGeometry args={[min_align_radius, 40]} />
-        <meshBasicMaterial color="rgb(0, 100, 100)" opacity={0.25} transparent={true} />
-      </mesh>
-      <mesh position={[rDFf?.x, rDFf?.y, 0.006]}>
-        <circleBufferGeometry args={[max_align_radius, 40]} />
-        <meshBasicMaterial color="rgb(0, 100, 100)" opacity={0.25} transparent={true} />
-      </mesh>
-      <mesh position={[rTFf?.x, rTFf?.y, 0.007]} rotation={[0, 0, target_rotation.z - 0.5 * min_angle_error]}>
-        <circleBufferGeometry args={[max_align_radius, 40, 0, min_angle_error]} />
-        <meshBasicMaterial color="rgb(0, 100, 100)" opacity={0.25} transparent={true} />
-      </mesh>
-      <mesh position={[rTFf?.x, rTFf?.y, 0.008]} rotation={[0, 0, target_rotation.z - 0.5 * max_angle_error]}>
-        <circleBufferGeometry args={[max_align_radius, 40, 0, max_angle_error]} />
-        <meshBasicMaterial color="rgb(0, 100, 100)" opacity={0.25} transparent={true} />
-      </mesh>
-      <mesh position={[rTFf?.x, rTFf?.y, 0.009]}>
-        <mesh geometry={arrowGeometry(max_align_radius)} rotation={[0, 0, robot_rotation.z]}>
-          <meshBasicMaterial color="rgb(255, 255, 255)" opacity={0.5} transparent={true} />
-        </mesh>
-      </mesh>
-      <mesh position={[rTFf?.x, rTFf?.y, 0.011]}>
-        <mesh geometry={arrowGeometry(speed)} rotation={[0, 0, velocity_direction]}>
-          <meshBasicMaterial color="rgb(0, 255, 0)" opacity={0.5} transparent={true} />
-        </mesh>
-      </mesh>
-      <mesh position={[rDFf?.x, rDFf?.y, 0.011]}>
-        <mesh geometry={arrowGeometry(min_align_radius)} rotation={[0, 0, robot_rotation.z + angle_to_final_heading]}>
-          <meshBasicMaterial color="rgb(255, 0, 0)" opacity={0.5} transparent={true} />
-        </mesh>
-      </mesh>
-    </object3D>
-  );
-};
-
-const BoundingBox = ({ model }: { model: LocalisationRobotModel }) => {
-  if (!model.boundingBox) return null;
-  const { minX, maxX, minY, maxY } = model.boundingBox;
-  const width = maxX - minX;
-  const height = maxY - minY;
-  const centerX = (minX + maxX) / 2;
-  const centerY = (minY + maxY) / 2;
-  const wallThickness = 0.05;
-  const wallHeight = 0.25;
-  const solidBottomHeight = 0.001;
-
-  const vertexShader = `
-    varying vec3 vPosition;
-    void main() {
-      vPosition = position;
-      gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
-    }
-  `;
-
-  const fragmentShader = `
-    uniform vec3 color;
-    uniform float wallHeight;
-    uniform float solidBottomHeight;
-    uniform int orientation;
-    varying vec3 vPosition;
-    void main() {
-      float heightFactor;
-      if (orientation == 0) {
-        heightFactor = vPosition.y / wallHeight + 0.5;
-      } else {
-        heightFactor = vPosition.z / wallHeight + 0.5;
-      }
-
-      float normalizedSolidHeight = solidBottomHeight / wallHeight;
-      float opacity;
-
-      if (heightFactor < normalizedSolidHeight) {
-        opacity = 0.9;
-      } else {
-        opacity = smoothstep(1.0, normalizedSolidHeight, heightFactor) * 0.3;
-      }
-
-      gl_FragColor = vec4(color, opacity);
-    }
-  `;
-
-  const createWallMaterial = (orientation: number) => {
-    return new THREE.ShaderMaterial({
-      uniforms: {
-        color: { value: new THREE.Color(model.color) },
-        wallHeight: { value: wallHeight },
-        solidBottomHeight: { value: solidBottomHeight },
-        orientation: { value: orientation },
-      },
-      vertexShader: vertexShader,
-      fragmentShader: fragmentShader,
-      transparent: true,
-      side: THREE.DoubleSide,
-    });
-  };
-
-  const verticalWallMaterial = useMemo(() => createWallMaterial(0), []);
-  const horizontalWallMaterial = useMemo(() => createWallMaterial(1), []);
-
-  // Update material color when model.color changes
-  useEffect(() => {
-    const newColor = new THREE.Color(model.color);
-    verticalWallMaterial.uniforms.color.value = newColor;
-    horizontalWallMaterial.uniforms.color.value = newColor;
-  }, [model.color, verticalWallMaterial, horizontalWallMaterial]);
-
-  return (
-    <object3D position={[centerX, centerY, wallHeight / 2]}>
-      {/* Left wall */}
-      <mesh position={[-(width / 2 + wallThickness / 2), 0, 0.009]} rotation={[Math.PI / 2, Math.PI / 2, 0]}>
-        <boxGeometry args={[height + wallThickness, wallHeight, wallThickness]} />
-        <primitive object={verticalWallMaterial} />
-      </mesh>
-      {/* Right wall */}
-      <mesh position={[width / 2 + wallThickness / 2, 0, 0.009]} rotation={[Math.PI / 2, Math.PI / 2, 0]}>
-        <boxGeometry args={[height + wallThickness, wallHeight, wallThickness]} />
-        <primitive object={verticalWallMaterial} />
-      </mesh>
-      {/* Top wall */}
-      <mesh position={[0, height / 2 + wallThickness / 2, 0.009]} rotation={[0, 0, 0]}>
-        <boxGeometry args={[width + wallThickness * 2, wallThickness, wallHeight]} />
-        <primitive object={horizontalWallMaterial} />
-      </mesh>
-      {/* Bottom wall */}
-      <mesh position={[0, -(height / 2 + wallThickness / 2), 0.009]} rotation={[0, 0, 0]}>
-        <boxGeometry args={[width + wallThickness * 2, wallThickness, wallHeight]} />
-        <primitive object={horizontalWallMaterial} />
-      </mesh>
-    </object3D>
-  );
-};
-
-const PurposeLabel = ({
-  robotModel,
-  cameraPitch,
-  cameraYaw,
-}: {
-  robotModel: LocalisationRobotModel;
-  cameraPitch: number;
-  cameraYaw: number;
-}) => {
-  const rTFf = robotModel.Hft.decompose().translation;
-  const textGeometry = (x: string) => {
-    const font = new FontLoader().parse(roboto);
-    return new TextGeometry(x, {
-      font: font,
-      size: 0.1,
-      height: 0,
-    }).center();
-  };
-
-  const textBackdropGeometry = (width: number, height: number) => {
-    const shape = new THREE.Shape();
-    width += 0.1;
-    height += 0.1;
-    const radius = 0.05;
-    const x = width * -0.5;
-    const y = height * -0.5;
-
-    shape.moveTo(x, y + radius);
-    shape.lineTo(x, y + height - radius);
-    shape.quadraticCurveTo(x, y + height, x + radius, y + height);
-    shape.lineTo(x + width - radius, y + height);
-    shape.quadraticCurveTo(x + width, y + height, x + width, y + height - radius);
-    shape.lineTo(x + width, y + radius);
-    shape.quadraticCurveTo(x + width, y, x + width - radius, y);
-    shape.lineTo(x + radius, y);
-    shape.quadraticCurveTo(x, y, x, y + radius);
-
-    const geometry = new THREE.ShapeGeometry(shape);
-
-    return geometry;
-  };
-
-  const label = robotModel.player_id == -1 ? robotModel.purpose : "N" + robotModel.player_id + " " + robotModel.purpose;
-  const labelTextGeometry = textGeometry(label);
-  labelTextGeometry.computeBoundingBox();
-  const textWidth = labelTextGeometry.boundingBox
-    ? labelTextGeometry.boundingBox.max.x - labelTextGeometry.boundingBox.min.x
-    : 0;
-  const textHeight = labelTextGeometry.boundingBox
-    ? labelTextGeometry.boundingBox.max.y - labelTextGeometry.boundingBox.min.y
-    : 0;
-  const backdropGeometry = textBackdropGeometry(textWidth, textHeight);
-
-  return (
-    <object3D
-      position={[rTFf?.x, rTFf?.y, rTFf?.z + 0.6]}
-      rotation={[Math.PI / 2 + cameraPitch, 0, -Math.PI / 2 + cameraYaw, "ZXY"]}
-    >
-      <mesh position={[0, 0, 0.001]} geometry={labelTextGeometry}>
-        <meshBasicMaterial color="white" transparent opacity={1} />
-      </mesh>
-      <mesh geometry={backdropGeometry}>
-        <meshBasicMaterial color={robotModel.color} transparent opacity={0.5} />
-      </mesh>
-    </object3D>
-  );
-};
-
-const FieldLinePoints = ({ model }: { model: LocalisationModel }) => (
-  <>
-    {model.robots.map(
-      (robot) =>
-        robot.visible && (
-          <object3D key={robot.id}>
-            {robot.rPFf.map((d, i) => {
-              return (
-                <mesh key={String(i)} position={d.add(new Vector3(0, 0, 0.005)).toArray()}>
-                  <circleBufferGeometry args={[0.02, 20]} />
-                  <meshBasicMaterial color="blue" />
-                </mesh>
-              );
-            })}
-          </object3D>
-        ),
-    )}
-  </>
-);
-
-const FieldIntersections = ({ model }: { model: LocalisationModel }) => {
-  return (
-    <>
-      {model.robots.map(
-        (robot) =>
-          robot.visible && (
-            <object3D key={robot.id}>
-              {robot.fieldIntersectionsF?.map((intersection) => {
-                const createShapeForIntersection = (intersectionType: string, position: Vector3) => {
-                  const basePosition = position.add(new Vector3(0.1, 0.1, 0)).toArray();
-                  switch (intersectionType) {
-                    case "L_INTERSECTION":
-                      return (
-                        <>
-                          <mesh position={intersection.position.add(new Vector3(0, 0, 0.01)).toArray()}>
-                            <circleBufferGeometry args={[0.04, 20]} />
-                            <meshBasicMaterial color="red" />
-                          </mesh>
-                          <mesh position={[basePosition[0], basePosition[1] - 0.05, basePosition[2]]}>
-                            <boxBufferGeometry args={[0.1, 0.02, 0.02]} />
-                            <meshBasicMaterial color="black" />
-                          </mesh>
-                          <mesh position={[basePosition[0] - 0.04, basePosition[1], basePosition[2]]}>
-                            <boxBufferGeometry args={[0.02, 0.1, 0.02]} />
-                            <meshBasicMaterial color="black" />
-                          </mesh>
-                        </>
-                      );
-                    case "T_INTERSECTION":
-                      return (
-                        <>
-                          <mesh position={intersection.position.add(new Vector3(0, 0, 0.01)).toArray()}>
-                            <circleBufferGeometry args={[0.04, 20]} />
-                            <meshBasicMaterial color="red" />
-                          </mesh>
-                          <mesh position={[basePosition[0], basePosition[1] + 0.05, basePosition[2]]}>
-                            <boxBufferGeometry args={[0.1, 0.02, 0.02]} />
-                            <meshBasicMaterial color="black" />
-                          </mesh>
-                          <mesh position={[basePosition[0], basePosition[1], basePosition[2]]}>
-                            <boxBufferGeometry args={[0.02, 0.1, 0.02]} />
-                            <meshBasicMaterial color="black" />
-                          </mesh>
-                        </>
-                      );
-                    case "X_INTERSECTION":
-                      return (
-                        <>
-                          <mesh position={intersection.position.add(new Vector3(0, 0, 0.01)).toArray()}>
-                            <circleBufferGeometry args={[0.04, 20]} />
-                            <meshBasicMaterial color="red" />
-                          </mesh>
-                          <mesh
-                            position={[basePosition[0], basePosition[1], basePosition[2]]}
-                            rotation={[0, 0, Math.PI / 4]}
-                          >
-                            <boxBufferGeometry args={[0.1, 0.02, 0.02]} />
-                            <meshBasicMaterial color="black" />
-                          </mesh>
-                          <mesh
-                            position={[basePosition[0], basePosition[1], basePosition[2]]}
-                            rotation={[0, 0, -Math.PI / 4]}
-                          >
-                            <boxBufferGeometry args={[0.1, 0.02, 0.02]} />
-                            <meshBasicMaterial color="black" />
-                          </mesh>
-                        </>
-                      );
-                    default:
-                      return null;
-                  }
-                };
-                return createShapeForIntersection(intersection.type, intersection.position);
-              })}
-            </object3D>
-          ),
-      )}
-    </>
-  );
-};
-
-const Particles = ({ model }: { model: LocalisationModel }) => (
-  <>
-    {model.robots.map(
-      (robot) =>
-        robot.visible && (
-          <object3D key={robot.id}>
-            {robot.particles.particle.map((particle, i) => {
-              return (
-                <mesh key={String(i)} position={new Vector3(particle.x, particle.y, 0.005).toArray()}>
-                  <circleBufferGeometry args={[0.02, 20]} />
-                  <meshBasicMaterial color="red" />
-                </mesh>
-              );
-            })}
-          </object3D>
-        ),
-    )}
-  </>
-);
-
-const Balls = ({ model }: { model: LocalisationModel }) => {
-  return (
-    <>
-      {model.robots.map(
-        (robot) =>
-          robot.visible &&
-          robot.rBFf && (
-            <mesh position={robot.rBFf.toArray()} scale={[robot.rBFf.z, robot.rBFf.z, robot.rBFf.z]} key={robot.id}>
-              <sphereBufferGeometry args={[1, 32, 32]} /> {/* Increased detail for the texture */}
-              <meshStandardMaterial map={soccerBallTexture} />
-            </mesh>
-          ),
-      )}
-    </>
-  );
-};
-
-const Goals = ({ model }: { model: LocalisationModel }) => (
-  <>
-    {model.robots.map(
-      (robot) =>
-        robot.visible &&
-        robot.rGFf && (
-          <object3D key={robot.id}>
-            {robot.rGFf.map((goal, i) => {
-              return (
-                <mesh
-                  key={String(i)}
-                  position={goal.bottom.add(new Vector3(0, 0, goal.top.z / 2)).toArray()}
-                  rotation={[Math.PI / 2, 0, 0]}
-                >
-                  <cylinderBufferGeometry args={[0.05, 0.05, goal.top.z, 20]} />
-                  <meshStandardMaterial color="magenta" />
-                </mesh>
-              );
-            })}
-          </object3D>
-        ),
-    )}
-  </>
-);
-
-const Robots = ({ model }: { model: LocalisationModel }) => (
-  <>
-    {model.robots.map(
-      (robot) =>
-        robot.visible &&
-        robot.robots && (
-          <object3D key={robot.id}>
-            {robot.rRFf.map((r, i) => {
-              return (
-                <mesh key={String(i)} position={r.add(new Vector3(0, 0, 0.4)).toArray()} rotation={[Math.PI / 2, 0, 0]}>
-                  <cylinderBufferGeometry args={[0.1, 0.1, 0.8, 20]} />
-                  <meshStandardMaterial color="orange" />
-                </mesh>
-              );
-            })}
-          </object3D>
-        ),
-    )}
-  </>
-);
-
-const WalkPathGoal = ({ model }: { model: LocalisationRobotModel }) => {
-  const robotRef = React.useRef<URDFRobot | null>(null);
-
-  // Load the URDF model only once
-  React.useEffect(() => {
-    const loader = new URDFLoader();
-    loader.load(nugusUrdfPath, (robot: URDFRobot) => {
-      if (robotRef.current) {
-        robotRef.current.add(robot);
-      }
-    });
-  }, []);
-
-  const rDFf = model.Hfd?.decompose().translation;
-  const rTFf = model.Hft.decompose().translation;
-  const Rfd_quat = new THREE.Quaternion(
-    model.Hfd?.decompose().rotation.x,
-    model.Hfd?.decompose().rotation.y,
-    model.Hfd?.decompose().rotation.z,
-    model.Hfd?.decompose().rotation.w,
-  );
-  const Rft_quat = new THREE.Quaternion(
-    model.Hft.decompose().rotation.x,
-    model.Hft.decompose().rotation.y,
-    model.Hft.decompose().rotation.z,
-    model.Hft.decompose().rotation.w,
-  );
-
-  // Get euler angles from quaternion
-  const Rfz_euler = new THREE.Euler().setFromQuaternion(Rfd_quat, "ZYX");
-  const Rft_euler = new THREE.Euler().setFromQuaternion(Rft_quat, "ZYX");
-  // Fuse the euler angles into a single quaternion
-  const rotation = new THREE.Quaternion().setFromEuler(new THREE.Euler(Rft_euler.x, Rft_euler.y, Rfz_euler.z, "ZYX"));
-  const position = new THREE.Vector3(rDFf?.x, rDFf?.y, rTFf.z);
-
-  const motors = model.motors;
-
-  // Update the position of the robot to match the walk path goal
-  React.useEffect(() => {
-    if (robotRef.current) {
-      robotRef.current.position.set(position.x, position.y, position.z);
-      robotRef.current.quaternion.copy(new THREE.Quaternion(rotation.x, rotation.y, rotation.z, rotation.w));
-      const joints = (robotRef.current?.children[0] as any)?.joints;
-      // Update robot's joints
-      if (joints) {
-        joints?.head_pitch.setJointValue(motors.headTilt.angle);
-        joints?.left_ankle_pitch.setJointValue(motors.leftAnklePitch.angle);
-        joints?.left_ankle_roll.setJointValue(motors.leftAnkleRoll.angle);
-        joints?.left_elbow_pitch.setJointValue(motors.leftElbow.angle);
-        joints?.left_hip_pitch.setJointValue(motors.leftHipPitch.angle);
-        joints?.left_hip_roll.setJointValue(motors.leftHipRoll.angle);
-        joints?.left_hip_yaw.setJointValue(motors.leftHipYaw.angle);
-        joints?.left_knee_pitch.setJointValue(motors.leftKnee.angle);
-        joints?.left_shoulder_pitch.setJointValue(motors.leftShoulderPitch.angle);
-        joints?.left_shoulder_roll.setJointValue(motors.leftShoulderRoll.angle);
-        joints?.neck_yaw.setJointValue(motors.headPan.angle);
-        joints?.right_ankle_pitch.setJointValue(motors.rightAnklePitch.angle);
-        joints?.right_ankle_roll.setJointValue(motors.rightAnkleRoll.angle);
-        joints?.right_elbow_pitch.setJointValue(motors.rightElbow.angle);
-        joints?.right_hip_pitch.setJointValue(motors.rightHipPitch.angle);
-        joints?.right_hip_roll.setJointValue(motors.rightHipRoll.angle);
-        joints?.right_hip_yaw.setJointValue(motors.rightHipYaw.angle);
-        joints?.right_knee_pitch.setJointValue(motors.rightKnee.angle);
-        joints?.right_shoulder_pitch.setJointValue(motors.rightShoulderPitch.angle);
-        joints?.right_shoulder_roll.setJointValue(motors.rightShoulderRoll.angle);
-      }
-      robotRef.current.traverse((child) => {
-        if (child instanceof THREE.Mesh) {
-          // Set opacity for all mesh children
-          child.material.transparent = true;
-          // Red
-          child.material.color = "rgb(0, 100, 100)";
-          child.material.opacity = 0.2;
-        }
-      });
-    }
-  });
-
-  return <object3D ref={robotRef} />;
-};
-
-const Robot = ({ model }: { model: LocalisationRobotModel }) => {
-  const robotRef = React.useRef<URDFRobot | null>(null);
-
-  // Load the URDF model only once
-  React.useEffect(() => {
-    const loader = new URDFLoader();
-    loader.load(nugusUrdfPath, (robot: URDFRobot) => {
-      if (robotRef.current) {
-        robotRef.current.add(robot);
-      }
-    });
-  }, [nugusUrdfPath]);
-
-  const position = model.Hft.decompose().translation;
-  const rotation = model.Hft.decompose().rotation;
-  const motors = model.motors;
-
-  React.useEffect(() => {
-    // Update robot's pose
-    if (robotRef.current) {
-      robotRef.current.position.copy(new THREE.Vector3(position.x, position.y, position.z));
-      robotRef.current.quaternion.copy(new THREE.Quaternion(rotation.x, rotation.y, rotation.z, rotation.w));
-      const joints = (robotRef.current?.children[0] as any)?.joints;
-      // Update robot's joints
-      if (joints) {
-        joints?.head_pitch.setJointValue(motors.headTilt.angle);
-        joints?.left_ankle_pitch.setJointValue(motors.leftAnklePitch.angle);
-        joints?.left_ankle_roll.setJointValue(motors.leftAnkleRoll.angle);
-        joints?.left_elbow_pitch.setJointValue(motors.leftElbow.angle);
-        joints?.left_hip_pitch.setJointValue(motors.leftHipPitch.angle);
-        joints?.left_hip_roll.setJointValue(motors.leftHipRoll.angle);
-        joints?.left_hip_yaw.setJointValue(motors.leftHipYaw.angle);
-        joints?.left_knee_pitch.setJointValue(motors.leftKnee.angle);
-        joints?.left_shoulder_pitch.setJointValue(motors.leftShoulderPitch.angle);
-        joints?.left_shoulder_roll.setJointValue(motors.leftShoulderRoll.angle);
-        joints?.neck_yaw.setJointValue(motors.headPan.angle);
-        joints?.right_ankle_pitch.setJointValue(motors.rightAnklePitch.angle);
-        joints?.right_ankle_roll.setJointValue(motors.rightAnkleRoll.angle);
-        joints?.right_elbow_pitch.setJointValue(motors.rightElbow.angle);
-        joints?.right_hip_pitch.setJointValue(motors.rightHipPitch.angle);
-        joints?.right_hip_roll.setJointValue(motors.rightHipRoll.angle);
-        joints?.right_hip_yaw.setJointValue(motors.rightHipYaw.angle);
-        joints?.right_knee_pitch.setJointValue(motors.rightKnee.angle);
-        joints?.right_shoulder_pitch.setJointValue(motors.rightShoulderPitch.angle);
-        joints?.right_shoulder_roll.setJointValue(motors.rightShoulderRoll.angle);
-      }
-    }
-  }, [position, rotation, motors]);
-
-  // Update the material of the robot
-  const material = new THREE.MeshStandardMaterial({
-    color: "#666666",
-    roughness: 0.5,
-    metalness: 0.2,
-  });
-  if (robotRef.current) {
-    robotRef.current.traverse((child) => {
-      if (child.type === "URDFVisual" && child.children.length > 0) {
-        const mesh = child.children[0] as THREE.Mesh;
-        mesh.material = material;
-      }
-    });
-  }
-
-  return <object3D ref={robotRef} />;
-};
